Refactor flags extension methods out of native methods
